move approx_top_k tests to self suite
